@@ -1,6 +1,5 @@
 ﻿using AdventOfCode.Core;
 using AdventOfCode.Core.IO;
-using AdventOfCode.Lib.Collections;
 using AdventOfCode.Lib.Math;
 
 namespace AdventOfCode2022.Challenges;
@@ -23,8 +22,8 @@ public class Challenge20
 
         for (var i = 0; i < numbers.Count; i++)
         {
-            int idx = indices.IndexOf(i);
-            int num = numbers[idx];
+            var idx = indices.IndexOf(i);
+            var num = numbers[idx];
             numbers.RemoveAt(idx);
             indices.RemoveAt(idx);
             var newIndex = Euclid.Modulus(idx + num, numbers.Count);
@@ -35,15 +34,17 @@ public class Challenge20
             }
             else
             {
-                numbers.Insert( newIndex, num);
+                numbers.Insert(newIndex, num);
                 indices.Insert(newIndex, i);
             }
         }
 
         var idxOf0 = numbers.IndexOf(0);
-        return (numbers[Euclid.Modulus(idxOf0 + 1000, numbers.Count)] + numbers[Euclid.Modulus(idxOf0 + 2000, numbers.Count)] + numbers[Euclid.Modulus(idxOf0 + 3000, numbers.Count)]).ToString();
+        return (numbers[Euclid.Modulus(idxOf0 + 1000, numbers.Count)] +
+                numbers[Euclid.Modulus(idxOf0 + 2000, numbers.Count)] +
+                numbers[Euclid.Modulus(idxOf0 + 3000, numbers.Count)]).ToString();
     }
-    
+
     [Part2]
     public async Task<string> Part2Async()
     {
@@ -51,28 +52,28 @@ public class Challenge20
         var indices = Enumerable.Range(0, numbers.Count).ToList();
 
         for (var j = 0; j < 10; j++)
+        for (var i = 0; i < numbers.Count; i++)
         {
-            for (var i = 0; i < numbers.Count; i++)
+            var idx = indices.IndexOf(i);
+            var num = numbers[idx];
+            numbers.RemoveAt(idx);
+            indices.RemoveAt(idx);
+            var newIndex = (int) Euclid.Modulus(idx + num, numbers.Count);
+            if (newIndex == 0)
             {
-                int idx = indices.IndexOf(i);
-                long num = numbers[idx];
-                numbers.RemoveAt(idx);
-                indices.RemoveAt(idx);
-                var newIndex = (int) Euclid.Modulus(idx + num, numbers.Count);
-                if (newIndex == 0)
-                {
-                    numbers.Add(num);
-                    indices.Add(i);
-                }
-                else
-                {
-                    numbers.Insert(newIndex, num);
-                    indices.Insert(newIndex, i);
-                }
+                numbers.Add(num);
+                indices.Add(i);
+            }
+            else
+            {
+                numbers.Insert(newIndex, num);
+                indices.Insert(newIndex, i);
             }
         }
 
         var idxOf0 = numbers.IndexOf(0);
-        return (numbers[Euclid.Modulus(idxOf0 + 1000, numbers.Count)] + numbers[Euclid.Modulus(idxOf0 + 2000, numbers.Count)] + numbers[Euclid.Modulus(idxOf0 + 3000, numbers.Count)]).ToString();
+        return (numbers[Euclid.Modulus(idxOf0 + 1000, numbers.Count)] +
+                numbers[Euclid.Modulus(idxOf0 + 2000, numbers.Count)] +
+                numbers[Euclid.Modulus(idxOf0 + 3000, numbers.Count)]).ToString();
     }
 }
