@@ -3,10 +3,6 @@ using AdventOfCode.Core.IO;
 using AdventOfCode.Lib;
 using AdventOfCode.Lib.Math;
 using AdventOfCode.Lib.PathFinding;
-using System.Collections;
-using System.Data;
-using System.Diagnostics;
-using System.Text;
 
 namespace AdventOfCode2022.Challenges;
 
@@ -25,18 +21,18 @@ public class Challenge24
     {
         var grid = await _inputReader.ReadGridAsync(24);
 
-        Point2 start = new Point2(1, 0);
-        Point2 goal = new Point2(120, 26);
+        var start = new Point2(1, 0);
+        var goal = new Point2(120, 26);
 
         var state = new State(start, 0);
 
-        var weight = new Func<State, State, int>((c, n) => 1);
+        var weight = new Func<State, State, int>((_, _) => 1);
         var heuristic = new Func<State, int>(c => Point2.ManhattanDistance(c.Position, goal));
         var neighbors = new Func<State, IEnumerable<State>>(c => GetAdjecent(grid, c));
         var goalReached = new Func<State, bool>(c => c.Position == goal);
 
         var astar = new AStar<State>(neighbors, weight, heuristic);
-        astar.TryPath(state, goalReached, out var path, out var minutes);
+        astar.TryPath(state, goalReached, out _, out var minutes);
 
         return minutes.ToString();
     }
@@ -47,12 +43,12 @@ public class Challenge24
     {
         var grid = await _inputReader.ReadGridAsync(24);
 
-        Point2 start = new Point2(1, 0);
-        Point2 goal = new Point2(120, 26);
+        var start = new Point2(1, 0);
+        var goal = new Point2(120, 26);
 
         var state = new State2(start, 0, 0);
 
-        var weight = new Func<State2, State2, int>((c, n) => 1);
+        var weight = new Func<State2, State2, int>((_, _) => 1);
         var heuristic = new Func<State2, int>(c =>
         {
             if (c.Phase == 1)
@@ -65,7 +61,7 @@ public class Challenge24
         var goalReached = new Func<State2, bool>(c => c.Position == goal && c.Phase == 2);
 
         var astar = new AStar<State2>(neighbors, weight, heuristic);
-        astar.TryPath(state, goalReached, out var path, out var minutes);
+        astar.TryPath(state, goalReached, out _, out var minutes);
 
         return minutes.ToString();
     }
@@ -74,7 +70,8 @@ public class Challenge24
     {
         foreach (var neighbor in current.Position.GetNeighbors())
         {
-            if (neighbor.X < 0 || neighbor.X >= grid.GetLength(1) || neighbor.Y < 0 || neighbor.Y >= grid.GetLength(0) || grid[neighbor.Y, neighbor.X] == '#')
+            if (neighbor.X < 0 || neighbor.X >= grid.GetLength(1) || neighbor.Y < 0 ||
+                neighbor.Y >= grid.GetLength(0) || grid[neighbor.Y, neighbor.X] == '#')
                 continue;
 
             if (!HasBlizzard(grid, neighbor, current.Time + 1))
@@ -95,7 +92,8 @@ public class Challenge24
 
         foreach (var neighbor in current.Position.GetNeighbors())
         {
-            if (neighbor.X < 0 || neighbor.X >= grid.GetLength(1) || neighbor.Y < 0 || neighbor.Y >= grid.GetLength(0) || grid[neighbor.Y, neighbor.X] == '#')
+            if (neighbor.X < 0 || neighbor.X >= grid.GetLength(1) || neighbor.Y < 0 ||
+                neighbor.Y >= grid.GetLength(0) || grid[neighbor.Y, neighbor.X] == '#')
                 continue;
 
             if (!HasBlizzard(grid, neighbor, current.Time + 1))
@@ -109,12 +107,15 @@ public class Challenge24
     private static bool HasBlizzard(char[,] grid, Point2 p, int time)
     {
         return grid[Wrap(p.Y - time, grid.GetLength(0)), p.X] == 'v'
-            || grid[Wrap(p.Y + time, grid.GetLength(0)), p.X] == '^'
-            || grid[p.Y, Wrap(p.X - time, grid.GetLength(1))] == '>'
-            || grid[p.Y, Wrap(p.X + time, grid.GetLength(1))] == '<';
+               || grid[Wrap(p.Y + time, grid.GetLength(0)), p.X] == '^'
+               || grid[p.Y, Wrap(p.X - time, grid.GetLength(1))] == '>'
+               || grid[p.Y, Wrap(p.X + time, grid.GetLength(1))] == '<';
     }
 
-    private static int Wrap(int v, int max) => Euclid.Modulus(v - 1, max - 2) + 1;
+    private static int Wrap(int v, int max)
+    {
+        return Euclid.Modulus(v - 1, max - 2) + 1;
+    }
 
     public record State(Point2 Position, int Time);
 
